Factor out stats calculation
