+++ src/sequali/__init__.py
@@ -33,11 +33,11 @@ from .plots import (adapter_content_plot, base_content_plot,
                     per_sequence_quality_scores_plot,
                     per_tile_graph,
                     sequence_length_distribution_plot)
-
-PHRED_TO_ERROR_RATE = [
-    sum(10 ** (-p / 10) for p in range(start * 4, start * 4 + 4)) / 4
-    for start in range(NUMBER_OF_PHREDS)
-]
+from .stats import (base_content, base_weighted_categories, equidistant_ranges,
+                    adapter_counts, per_base_qualities, mean_qualities,
+                    aggregate_sequence_lengths, min_length, q20_bases,
+                    total_gc_fraction, aggregate_count_matrix,
+                    stringify_ranges, normalized_per_tile_averages, )
 
 __all__ = [
     "A", "C", "G", "N", "T",
@@ -54,325 +54,59 @@ __all__ = [
 ]
 
 
-def equidistant_ranges(length: int, parts: int) -> Iterator[Tuple[int, int]]:
-    size = length // parts
-    remainder = length % parts
-    small_parts = parts - remainder
-    start = 0
-    for i in range(parts):
-        part_size = size if i < small_parts else size + 1
-        if part_size == 0:
-            continue
-        stop = start + part_size
-        yield start, stop
-        start = stop
-
-
-def base_weighted_categories(
-        base_counts: Sequence[int], number_of_categories: int
-) -> Iterator[Tuple[int, int]]:
-    total_bases = sum(base_counts)
-    per_category = total_bases // number_of_categories
-    enough_bases = per_category
-    start = 0
-    total = 0
-    for stop, count in enumerate(base_counts, start=1):
-        total += count
-        if total >= enough_bases:
-            yield start, stop
-            start = stop
-            enough_bases += per_category
-    if start != len(base_counts):
-        yield start, len(base_counts)
-
-
-def cumulative_percentages(counts: Iterable[int], total: int):
-    cumalitive_percentages = []
-    count_sum = 0
-    for count in counts:
-        count_sum += count
-        cumalitive_percentages.append(count_sum / total)
-    return cumalitive_percentages
-
-
-def normalized_per_tile_averages(
-        per_tile_quality: PerTileQuality) -> List[Tuple[str, List[float]]]:
-    tile_averages = per_tile_quality.get_tile_averages()
-    max_length = per_tile_quality.max_length
-    ranges = list(equidistant_ranges(max_length, 50))
-
-    average_phreds = []
-    per_category_totals = [0.0 for i in range(len(ranges))]
-    for tile, averages in tile_averages:
-        range_averages = [sum(averages[start:stop]) / (stop - start)
-                          for start, stop in ranges]
-        range_phreds = []
-        for i, average in enumerate(range_averages):
-            phred = -10 * math.log10(average)
-            range_phreds.append(phred)
-            # Averaging phreds takes geometric mean.
-            per_category_totals[i] += phred
-        average_phreds.append((tile, range_phreds))
-    number_of_tiles = len(tile_averages)
-    averages_per_category = [total / number_of_tiles
-                             for total in per_category_totals]
-    normalized_averages = []
-    for tile, tile_phreds in average_phreds:
-        normalized_tile_phreds = [
-            tile_phred - average
-            for tile_phred, average in zip(tile_phreds, averages_per_category)
-        ]
-        normalized_averages.append((str(tile), normalized_tile_phreds))
-    return normalized_averages
-
-
-class QCMetricsReport:
-    raw_count_matrix: array.ArrayType
-    aggregated_count_matrix: array.ArrayType
-    raw_sequence_lengths: array.ArrayType
-    gc_content: array.ArrayType
-    phred_scores: array.ArrayType
-    _data_ranges: List[Tuple[int, int]]
-    data_categories: List[str]
-    max_length: int
-    total_reads: int
-    total_bases: int
-
-    def __init__(self, metrics: QCMetrics, adapter_counter: AdapterCounter,
-                 graph_resolution: int = 100):
-        """Aggregate all data from a QCMetrics counter"""
-
-        self.adapter_counter = adapter_counter
-        self.total_reads = metrics.number_of_reads
-        assert metrics.number_of_reads == adapter_counter.number_of_sequences
-        self.max_length = metrics.max_length
-        self.raw_count_matrix = array.array("Q")
-        # Python will treat the memoryview as an iterable in the array constructor
-        # use from_bytes instead for direct memcpy.
-        self.raw_count_matrix = metrics.count_table()
-        self.gc_content = metrics.gc_content()
-        self.phred_scores = metrics.phred_scores()
-
-        matrix = memoryview(self.raw_count_matrix)
-
-        # use bytes constructor to initialize the aggregated count matrix to 0.
-        raw_sequence_lengths = array.array("Q",
-                                           bytes(8 * (self.max_length + 1)))
-        raw_base_counts = array.array("Q", bytes(8 * (self.max_length + 1)))
-        # All reads have at least 0 bases
-        raw_base_counts[0] = self.total_reads
-        for i in range(self.max_length):
-            table = matrix[i * 60:(i + 1) * 60]
-            raw_base_counts[i + 1] = sum(table)
-
-        previous_count = 0
-        for i in range(self.max_length, 0, -1):
-            number_at_least = raw_base_counts[i]
-            raw_sequence_lengths[i] = number_at_least - previous_count
-            previous_count = number_at_least
-        self.raw_sequence_lengths = raw_sequence_lengths
-        self.total_bases = sum(memoryview(raw_base_counts)[1:])
-
-        expected_bases_per_position = self.total_bases / self.max_length
-        if raw_base_counts[-1] > (0.5 * expected_bases_per_position):
-            # Most reads are max length. Use an equidistant distribution
-            self._data_ranges = list(
-                equidistant_ranges(metrics.max_length, graph_resolution)
-            )
-        else:
-            # We have greater variance in the length of the reads. Use the
-            # base counts to get roughly equal base counts per category.
-            self._data_ranges = list(
-                base_weighted_categories(
-                    memoryview(raw_base_counts)[1:], graph_resolution
-                )
-            )
-
-        # Use one-based indexing for the graph categories. I.e. 1 is the first base.
-        self.data_categories = [
-            f"{start + 1}-{stop}" if start + 1 != stop else f"{start + 1}"
-            for start, stop in self._data_ranges
-        ]
-
-        self.aggregated_count_matrix = array.array(
-            "Q", bytes(8 * TABLE_SIZE * len(self.data_categories))
-        )
-        categories_view = memoryview(self.aggregated_count_matrix)
-        table_size = TABLE_SIZE
-        for cat_index, (start, stop) in enumerate(self._data_ranges):
-            cat_offset = cat_index * table_size
-            cat_view = categories_view[cat_offset: cat_offset + table_size]
-            for table_index in range(start, stop):
-                offset = table_index * table_size
-                table = matrix[offset: offset + table_size]
-                for i, count in enumerate(table):
-                    cat_view[i] += count
-
-    def _tables(self) -> Iterator[memoryview]:
-        category_view = memoryview(self.aggregated_count_matrix)
-        for i in range(0, len(category_view), TABLE_SIZE):
-            yield category_view[i: i + TABLE_SIZE]
-
-    def base_content(self) -> List[List[float]]:
-        content = [
-            [0.0 for _ in range(len(self.data_categories))]
-            for _ in range(NUMBER_OF_NUCS)
-        ]
-        for cat_index, table in enumerate(self._tables()):
-            total = sum(table)
-            if total == 0:
-                continue
-            for i in range(NUMBER_OF_NUCS):
-                content[i][cat_index] = sum(table[i::NUMBER_OF_NUCS]) / total
-        return content
-
-    def total_gc_fraction(self) -> float:
-        total_nucs = [
-            sum(
-                self.aggregated_count_matrix[
-                    i: len(self.aggregated_count_matrix): NUMBER_OF_NUCS
-                ]
-            )
-            for i in range(NUMBER_OF_NUCS)
-        ]
-        at = total_nucs[A] + total_nucs[T]
-        gc = total_nucs[G] + total_nucs[C]
-        return gc / (at + gc)
-
-    def q20_bases(self):
-        q20s = 0
-        for table in self._tables():
-            q20s += sum(table[NUMBER_OF_NUCS * 5:])
-        return q20s
-
-    def q28_bases(self):
-        q28s = 0
-        for table in self._tables():
-            q28s += sum(table[NUMBER_OF_NUCS * 7:])
-        return q28s
-
-    def min_length(self) -> int:
-        for length, count in enumerate(self.raw_sequence_lengths):
-            if count > 0:
-                return length
-        return 0
-
-    def mean_length(self) -> float:
-        return self.total_bases / self.total_reads
-
-    def sequence_lengths(self):
-        seqlength_view = memoryview(self.raw_sequence_lengths)[1:]
-        lengths = [sum(seqlength_view[start:stop]) for start, stop in
-                   self._data_ranges]
-        return [self.raw_sequence_lengths[0]] + lengths
-
-    def mean_qualities(self):
-        mean_qualites = [0.0 for _ in range(len(self.data_categories))]
-        for cat_index, table in enumerate(self._tables()):
-            total = 0
-            total_prob = 0.0
-            for phred_p_value, offset in zip(
-                    PHRED_TO_ERROR_RATE, range(0, TABLE_SIZE, NUMBER_OF_NUCS)
-            ):
-                nucs = table[offset: offset + NUMBER_OF_NUCS]
-                count = sum(nucs)
-                total += count
-                total_prob += count * phred_p_value
-            if total == 0:
-                continue
-            mean_qualites[cat_index] = -10 * math.log10(total_prob / total)
-        return mean_qualites
-
-    def per_base_qualities(self) -> List[List[float]]:
-        base_qualities = [
-            [0.0 for _ in range(len(self.data_categories))]
-            for _ in range(NUMBER_OF_NUCS)
-        ]
-        for cat_index, table in enumerate(self._tables()):
-            nuc_probs = [0.0 for _ in range(NUMBER_OF_NUCS)]
-            nuc_counts = [0 for _ in range(NUMBER_OF_NUCS)]
-            for phred_p_value, offset in zip(
-                    PHRED_TO_ERROR_RATE, range(0, TABLE_SIZE, NUMBER_OF_NUCS)
-            ):
-                nucs = table[offset: offset + NUMBER_OF_NUCS]
-                for i, count in enumerate(nucs):
-                    nuc_counts[i] += count
-                    nuc_probs[i] += phred_p_value * count
-            for i in range(NUMBER_OF_NUCS):
-                if nuc_counts[i] == 0:
-                    continue
-                base_qualities[i][cat_index] = -10 * math.log10(
-                    nuc_probs[i] / nuc_counts[i]
-                )
-        return base_qualities
-
-    def all_adapter_values(self):
-        all_adapters = []
-        for adapter, countarray in self.adapter_counter.get_counts():
-            adapter_counts = [sum(countarray[start:stop])
-                              for start, stop in self._data_ranges]
-            total = 0
-            accumulated_counts = []
-            for count in adapter_counts:
-                total += count
-                accumulated_counts.append(total)
-            all_adapters.append([count * 100 / self.total_reads for
-                                 count in accumulated_counts])
-        return all_adapters
-
-    def html_report(self):
-        return f"""
-        <html>
-        <head>
-            <meta http-equiv="content-type" content="text/html:charset=utf-8">
-            <title>sequali report</title>
-        </head>
-        <h1>sequali report</h1>
-        <h2>Summary</h2>
-        <table>
-        <tr><td>Mean length</td><td align="right">
-            {self.mean_length():.2f}</td></tr>
-        <tr><td>Length range (min-max)</td><td align="right">
-            {self.min_length()}-{self.max_length}</td></tr>
-        <tr><td>total reads</td><td align="right">{self.total_reads}</td></tr>
-        <tr><td>total bases</td><td align="right">{self.total_bases}</td></tr>
-        <tr>
-            <td>Q20 bases</td>
-            <td align="right">
-                {self.q20_bases()} ({self.q20_bases() * 100 / self.total_bases:.2f}%)
-            </td>
-        </tr>
-        <tr>
-            <td>Q28 bases</td>
-            <td align="right">
-                {self.q28_bases()} ({self.q28_bases() * 100 / self.total_bases:.2f}%)
-            </td>
-        </tr>
-        <tr><td>GC content</td><td align="right">
-            {self.total_gc_fraction() * 100:.2f}%
-        </td></tr>
-        </table>
-        <h2>Quality scores</h2>
-        {per_base_quality_plot(self.per_base_qualities(),
-                               self.mean_qualities(),
-                               self.data_categories)}
-        </html>
-        <h2>Sequence length distribution</h2>
-        {sequence_length_distribution_plot(
-            sequence_lengths=self.sequence_lengths(),
-            x_labels=["0"] + self.data_categories)}
-        <h2>Base content</h2>
-        {base_content_plot(self.base_content(), self.data_categories)}
-        <h2>Per sequence GC content</h2>
-        {per_sequence_gc_content_plot(self.gc_content)}
-        <h2>Per sequence quality scores</h2>
-        {per_sequence_quality_scores_plot(self.phred_scores)}
-        <h2>Adapter content plot</h2>
-        {adapter_content_plot(self.all_adapter_values(),
-                              self.adapter_counter.adapters,
-                              self.data_categories)}
-        </html>
-        """
+def html_report(self):
+    return f"""
+    <html>
+    <head>
+        <meta http-equiv="content-type" content="text/html:charset=utf-8">
+        <title>sequali report</title>
+    </head>
+    <h1>sequali report</h1>
+    <h2>Summary</h2>
+    <table>
+    <tr><td>Mean length</td><td align="right">
+        {self.mean_length():.2f}</td></tr>
+    <tr><td>Length range (min-max)</td><td align="right">
+        {self.min_length()}-{self.max_length}</td></tr>
+    <tr><td>total reads</td><td align="right">{self.total_reads}</td></tr>
+    <tr><td>total bases</td><td align="right">{self.total_bases}</td></tr>
+    <tr>
+        <td>Q20 bases</td>
+        <td align="right">
+            {self.q20_bases()} ({self.q20_bases() * 100 / self.total_bases:.2f}%)
+        </td>
+    </tr>
+    <tr>
+        <td>Q28 bases</td>
+        <td align="right">
+            {self.q28_bases()} ({self.q28_bases() * 100 / self.total_bases:.2f}%)
+        </td>
+    </tr>
+    <tr><td>GC content</td><td align="right">
+        {self.total_gc_fraction() * 100:.2f}%
+    </td></tr>
+    </table>
+    <h2>Quality scores</h2>
+    {per_base_quality_plot(self.per_base_qualities(),
+                           self.mean_qualities(),
+                           self.data_categories)}
+    </html>
+    <h2>Sequence length distribution</h2>
+    {sequence_length_distribution_plot(
+        sequence_lengths=self.sequence_lengths(),
+        x_labels=["0"] + self.data_categories)}
+    <h2>Base content</h2>
+    {base_content_plot(self.base_content(), self.data_categories)}
+    <h2>Per sequence GC content</h2>
+    {per_sequence_gc_content_plot(self.gc_content)}
+    <h2>Per sequence quality scores</h2>
+    {per_sequence_quality_scores_plot(self.phred_scores)}
+    <h2>Adapter content plot</h2>
+    {adapter_content_plot(self.all_adapter_values(),
+                          self.adapter_counter.adapters,
+                          self.data_categories)}
+    </html>
+    """
 
 
 def main():
@@ -429,7 +163,8 @@ def main():
     print(
         per_tile_graph(
             normalized_per_tile_averages(per_tile_quality),
-            [f"{x}-{y}" for x, y in equidistant_ranges(per_tile_quality.max_length, 50)]
+            [f"{x}-{y}" for x, y in
+             equidistant_ranges(per_tile_quality.max_length, 50)]
         )
     )
