+++ src/sequali/stats.py
@@ -1,0 +1,258 @@
+# Copyright (C) 2023 Leiden University Medical Center
+# This file is part of sequali
+#
+# sequali is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# sequali is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with sequali.  If not, see <https://www.gnu.org/licenses/
+import array
+import math
+from typing import Iterable, Iterator, List, Sequence, Tuple
+
+from ._qc import A, C, G, N, T
+from ._qc import AdapterCounter, FastqParser, FastqRecordView, \
+    PerTileQuality, QCMetrics, SequenceDuplication
+from ._qc import NUMBER_OF_NUCS, NUMBER_OF_PHREDS, PHRED_MAX, TABLE_SIZE
+
+
+PHRED_TO_ERROR_RATE = [
+    sum(10 ** (-p / 10) for p in range(start * 4, start * 4 + 4)) / 4
+    for start in range(NUMBER_OF_PHREDS)
+]
+
+
+def equidistant_ranges(length: int, parts: int) -> Iterator[Tuple[int, int]]:
+    size = length // parts
+    remainder = length % parts
+    small_parts = parts - remainder
+    start = 0
+    for i in range(parts):
+        part_size = size if i < small_parts else size + 1
+        if part_size == 0:
+            continue
+        stop = start + part_size
+        yield start, stop
+        start = stop
+
+
+def base_weighted_categories(
+        base_counts: Sequence[int], number_of_categories: int
+) -> Iterator[Tuple[int, int]]:
+    total_bases = sum(base_counts)
+    per_category = total_bases // number_of_categories
+    enough_bases = per_category
+    start = 0
+    total = 0
+    for stop, count in enumerate(base_counts, start=1):
+        total += count
+        if total >= enough_bases:
+            yield start, stop
+            start = stop
+            enough_bases += per_category
+    if start != len(base_counts):
+        yield start, len(base_counts)
+
+
+def stringify_ranges(data_ranges: Iterable[Tuple[int, int]]):
+    return [
+            f"{start + 1}-{stop}" if start + 1 != stop else f"{start + 1}"
+            for start, stop in data_ranges
+    ]
+
+
+def cumulative_percentages(counts: Iterable[int], total: int):
+    cumalitive_percentages = []
+    count_sum = 0
+    for count in counts:
+        count_sum += count
+        cumalitive_percentages.append(count_sum / total)
+    return cumalitive_percentages
+
+
+def normalized_per_tile_averages(
+        per_tile_quality: PerTileQuality) -> List[Tuple[str, List[float]]]:
+    tile_averages = per_tile_quality.get_tile_averages()
+    max_length = per_tile_quality.max_length
+    ranges = list(equidistant_ranges(max_length, 50))
+
+    average_phreds = []
+    per_category_totals = [0.0 for i in range(len(ranges))]
+    for tile, averages in tile_averages:
+        range_averages = [sum(averages[start:stop]) / (stop - start)
+                          for start, stop in ranges]
+        range_phreds = []
+        for i, average in enumerate(range_averages):
+            phred = -10 * math.log10(average)
+            range_phreds.append(phred)
+            # Averaging phreds takes geometric mean.
+            per_category_totals[i] += phred
+        average_phreds.append((tile, range_phreds))
+    number_of_tiles = len(tile_averages)
+    averages_per_category = [total / number_of_tiles
+                             for total in per_category_totals]
+    normalized_averages = []
+    for tile, tile_phreds in average_phreds:
+        normalized_tile_phreds = [
+            tile_phred - average
+            for tile_phred, average in zip(tile_phreds, averages_per_category)
+        ]
+        normalized_averages.append((str(tile), normalized_tile_phreds))
+    return normalized_averages
+
+
+def table_iterator(count_tables: array.ArrayType[int]) -> Iterator[memoryview]:
+    table_view = memoryview(count_tables)
+    for i in range(0, len(count_tables), TABLE_SIZE):
+        yield table_view[i: i + TABLE_SIZE]
+
+
+def sequence_lengths(count_tables: array.ArrayType[int], total_reads: int):
+    max_length = len(count_tables) // TABLE_SIZE
+    # use bytes constructor to initialize to 0
+    sequence_lengths = array.array("Q", bytes(8 * (max_length + 1)))
+    base_counts = array.array("Q", bytes(8 * (max_length + 1)))
+    base_counts[0] = total_reads  # all reads have at least 0 bases
+    for i, table in enumerate(table_iterator(count_tables)):
+        base_counts[i+1] = sum(table)
+    previous_count = 0
+    for i in range(max_length, 0, -1):
+        number_at_least = base_counts[i]
+        sequence_lengths = number_at_least - previous_count
+    return sequence_lengths
+
+
+def aggregate_count_matrix(
+        count_tables: array.ArrayType[int],
+        data_ranges: Sequence[Tuple[int, int]]) -> array.ArrayType[int]:
+    count_view = memoryview(count_tables)
+    aggregated_matrix = array.array("Q", bytes(8 * TABLE_SIZE * len(data_ranges)))
+    ag_view = memoryview(aggregated_matrix)
+    for cat_index, (start, stop) in enumerate(data_ranges):
+        cat_offset = cat_index * TABLE_SIZE
+        cat_view = ag_view[cat_offset:cat_offset + TABLE_SIZE]
+        for table_index in range(start, stop):
+            offset = table_index * TABLE_SIZE
+            table = count_view[offset: offset + TABLE_SIZE]
+            for i, count in enumerate(table):
+                cat_view[i] += count
+    return aggregated_matrix
+
+
+def base_content(count_tables: array.ArrayType[int]) -> List[List[float]]:
+    total_tables = len(count_tables) // TABLE_SIZE
+    base_fractions = [
+        [0.0 for _ in range(total_tables)]
+        for _ in range(NUMBER_OF_NUCS)
+    ]
+    for index, table in enumerate(table_iterator(count_tables)):
+        total = sum(table)
+        if total == 0:
+            continue
+        for i in range(NUMBER_OF_NUCS):
+            base_fractions[i][index] = sum(table[i::NUMBER_OF_NUCS]) / total
+    return base_fractions
+
+
+def total_gc_fraction(count_tables: array.ArrayType[int]) -> float:
+    total_nucs = [
+        sum(
+            count_tables[
+            i: len(count_tables): NUMBER_OF_NUCS
+            ]
+        )
+        for i in range(NUMBER_OF_NUCS)
+    ]
+    at = total_nucs[A] + total_nucs[T]
+    gc = total_nucs[G] + total_nucs[C]
+    return gc / (at + gc)
+
+
+def q20_bases(count_tables: array.ArrayType[int]) -> int:
+    q20s = 0
+    for table in table_iterator(count_tables):
+        q20s += sum(table[NUMBER_OF_NUCS * 5:])
+    return q20s
+
+
+def min_length(sequence_lengths: Sequence[int]) -> int:
+    for length, count in enumerate(sequence_lengths):
+        if count > 0:
+            return length
+    return 0
+
+
+def aggregate_sequence_lengths(raw_sequence_lengths: array.ArrayType[int],
+                               data_ranges: Iterable[Tuple[int, int]]):
+    seqlength_view = memoryview(raw_sequence_lengths)[1:]
+    lengths = [sum(seqlength_view[start:stop]) for start, stop in
+               data_ranges]
+    return [raw_sequence_lengths[0]] + lengths
+
+
+def mean_qualities(count_tables: array.ArrayType[int]) -> List[float]:
+    total_tables = len(count_tables) // TABLE_SIZE
+    mean_qualities = [0.0 for _ in range(total_tables)]
+    for index, table in enumerate(table_iterator(count_tables)):
+        total = 0
+        total_prob = 0.0
+        for phred_p_value, offset in zip(
+                PHRED_TO_ERROR_RATE, range(0, TABLE_SIZE, NUMBER_OF_NUCS)
+        ):
+            nucs = table[offset: offset + NUMBER_OF_NUCS]
+            count = sum(nucs)
+            total += count
+            total_prob += count * phred_p_value
+        if total == 0:
+            continue
+        mean_qualities[index] = -10 * math.log10(total_prob / total)
+    return mean_qualities
+
+
+def per_base_qualities(count_tables: array.ArrayType[int]) -> List[List[float]]:
+    total_tables = len(count_tables) // TABLE_SIZE
+    base_qualities = [
+        [0.0 for _ in range(total_tables)]
+        for _ in range(NUMBER_OF_NUCS)
+    ]
+    for cat_index, table in enumerate(table_iterator(count_tables)):
+        nuc_probs = [0.0 for _ in range(NUMBER_OF_NUCS)]
+        nuc_counts = [0 for _ in range(NUMBER_OF_NUCS)]
+        for phred_p_value, offset in zip(
+                PHRED_TO_ERROR_RATE, range(0, TABLE_SIZE, NUMBER_OF_NUCS)
+        ):
+            nucs = table[offset: offset + NUMBER_OF_NUCS]
+            for i, count in enumerate(nucs):
+                nuc_counts[i] += count
+                nuc_probs[i] += phred_p_value * count
+        for i in range(NUMBER_OF_NUCS):
+            if nuc_counts[i] == 0:
+                continue
+            base_qualities[i][cat_index] = -10 * math.log10(
+                nuc_probs[i] / nuc_counts[i]
+            )
+    return base_qualities
+
+
+def adapter_counts(adapter_counter: AdapterCounter,
+                   data_ranges: Sequence[Tuple[int, int]]):
+    all_adapters = []
+    total_sequences = adapter_counter.number_of_sequences
+    for adapter, countarray in adapter_counter.get_counts():
+        adapter_counts = [sum(countarray[start:stop])
+                          for start, stop in data_ranges]
+        total = 0
+        accumulated_counts = []
+        for count in adapter_counts:
+            total += count
+            accumulated_counts.append(total)
+        all_adapters.append([count * 100 / total_sequences
+                             for count in accumulated_counts])
+    return all_adapters
